refactor : 커스텀 훅 관리를 위한 utils 디렉토리 생성
